Extract the Router APIs and context into a dedicated package
