+++ package.json
@@ -76,6 +76,7 @@
 		"@wordpress/redux-routine": "file:packages/redux-routine",
 		"@wordpress/reusable-blocks": "file:packages/reusable-blocks",
 		"@wordpress/rich-text": "file:packages/rich-text",
+		"@wordpress/router": "file:packages/router",
 		"@wordpress/server-side-render": "file:packages/server-side-render",
 		"@wordpress/shortcode": "file:packages/shortcode",
 		"@wordpress/style-engine": "file:packages/style-engine",
