+++ packages/edit-site/src/components/add-new-template/new-template.js
@@ -31,6 +31,7 @@ import {
 } from '@wordpress/icons';
 import { __, sprintf } from '@wordpress/i18n';
 import { store as noticesStore } from '@wordpress/notices';
+import { privateApis as routerPrivateApis } from '@wordpress/router';
 
 /**
  * Internal dependencies
@@ -46,9 +47,10 @@ import {
 } from './utils';
 import AddCustomGenericTemplateModal from './add-custom-generic-template-modal';
 import TemplateActionsLoadingScreen from './template-actions-loading-screen';
-import { useHistory } from '../routes';
 import { store as editSiteStore } from '../../store';
 import { unlock } from '../../private-apis';
+
+const { useHistory } = unlock( routerPrivateApis );
 
 const DEFAULT_TEMPLATE_SLUGS = [
 	'front-page',
