+++ packages/edit-site/src/components/sync-state-with-url/use-sync-canvas-mode-with-url.js
@@ -3,13 +3,15 @@
  */
 import { useEffect, useRef } from '@wordpress/element';
 import { useSelect, useDispatch } from '@wordpress/data';
+import { privateApis as routerPrivateApis } from '@wordpress/router';
 
 /**
  * Internal dependencies
  */
 import { store as editSiteStore } from '../../store';
-import { useLocation, useHistory } from '../routes';
 import { unlock } from '../../private-apis';
+
+const { useLocation, useHistory } = unlock( routerPrivateApis );
 
 export default function useSyncCanvasModeWithURL() {
 	const history = useHistory();
