+++ packages/private-apis/src/implementation.js
@@ -21,6 +21,7 @@ const CORE_MODULES_USING_PRIVATE_APIS = [
 	'@wordpress/edit-site',
 	'@wordpress/edit-widgets',
 	'@wordpress/editor',
+	'@wordpress/router',
 ];
 
 /**
